chore: move css.js into src
